+++ example/app/negative-bar-charts.tsx
@@ -1,13 +1,26 @@
 import { LinearGradient, useFont, vec } from "@shopify/react-native-skia";
 import React, { useState } from "react";
 import { SafeAreaView, ScrollView, StyleSheet, View } from "react-native";
-import { Bar, BarGroup, CartesianChart } from "victory-native";
+import { Bar, BarGroup, CartesianChart, StackedBar } from "victory-native";
 import { useDarkMode } from "react-native-dark";
 import inter from "../assets/inter-medium.ttf";
 import { appColors } from "./consts/colors";
 import { Button } from "../components/Button";
 import { InfoCard } from "../components/InfoCard";
+import { Text } from "../components/Text";
 import { descriptionForRoute } from "./consts/routes";
+import { Checkbox } from "../components/Checkbox";
+
+const STACKED_DATA = (length: number = 5) =>
+  Array.from({ length }, (_, index) => {
+    const y = Math.random() < 0.4 ? 0 : Math.floor(Math.random() * 9);
+    const z = Math.random() < 0.4 ? 0 : Math.floor(Math.random() * -9);
+    return {
+      x: index + 1,
+      y: y === 0 && z === 0 ? Math.floor(Math.random() * 9) : y,
+      z: z === 0 && y === 0 ? Math.floor(Math.random() * -9) : z,
+    };
+  });
 
 const GROUP_DATA = (length: number = 10) =>
   Array.from({ length }, (_, index) => ({
@@ -29,12 +42,19 @@ export default function NegativeBarChartsPage(props: { segment: string }) {
   const isDark = useDarkMode();
   const [data, setData] = useState(DATA(10));
   const [groupData, setGroupData] = useState(GROUP_DATA(5));
+  const [stackedData, setStackedData] = useState(STACKED_DATA(5));
   const [innerPadding] = useState(0.33);
+
   const [roundedCorner] = useState(5);
+  const [topLeft, setTopLeft] = useState(true);
+  const [topRight, setTopRight] = useState(true);
+  const [bottomRight, setBottomRight] = useState(false);
+  const [bottomLeft, setBottomLeft] = useState(false);
 
   function shuffleData() {
     setData(DATA(10));
     setGroupData(GROUP_DATA(5));
+    setStackedData(STACKED_DATA(5));
   }
 
   return (
@@ -62,8 +82,10 @@ export default function NegativeBarChartsPage(props: { segment: string }) {
                     animate={{ type: "spring" }}
                     innerPadding={innerPadding}
                     roundedCorners={{
-                      topLeft: roundedCorner,
-                      topRight: roundedCorner,
+                      topLeft: topLeft ? roundedCorner : 0,
+                      topRight: topRight ? roundedCorner : 0,
+                      bottomLeft: bottomLeft ? roundedCorner : 0,
+                      bottomRight: bottomRight ? roundedCorner : 0,
                     }}
                   >
                     <LinearGradient
@@ -96,8 +118,10 @@ export default function NegativeBarChartsPage(props: { segment: string }) {
                 betweenGroupPadding={0.4}
                 withinGroupPadding={0.1}
                 roundedCorners={{
-                  topLeft: roundedCorner,
-                  topRight: roundedCorner,
+                  topLeft: topLeft ? roundedCorner : 0,
+                  topRight: topRight ? roundedCorner : 0,
+                  bottomLeft: bottomLeft ? roundedCorner : 0,
+                  bottomRight: bottomRight ? roundedCorner : 0,
                 }}
               >
                 <BarGroup.Bar points={points.y} animate={{ type: "timing" }}>
@@ -125,6 +149,46 @@ export default function NegativeBarChartsPage(props: { segment: string }) {
             )}
           </CartesianChart>
         </View>
+        <View style={styles.chart}>
+          <CartesianChart
+            xKey="x"
+            yKeys={["y", "z"]}
+            padding={5}
+            domainPadding={{ left: 50, right: 50, top: 0 }}
+            domain={{ y: [-10, 10] }}
+            axisOptions={{
+              font,
+              formatXLabel: (value) => {
+                const date = new Date(2023, value - 1);
+                return date.toLocaleString("default", { month: "short" });
+              },
+              lineColor: isDark ? "#71717a" : "#d4d4d8",
+              labelColor: isDark ? appColors.text.dark : appColors.text.light,
+            }}
+            data={stackedData}
+          >
+            {({ points, chartBounds }) => {
+              return (
+                <StackedBar
+                  animate={{ type: "timing" }}
+                  barWidth={25}
+                  innerPadding={0.33}
+                  chartBounds={chartBounds}
+                  points={[points.y, points.z]}
+                  colors={["rgba(59,239,101,0.67)", "rgba(244,71,71,0.63)"]}
+                  barOptions={({ isBottom, isTop }) => ({
+                    roundedCorners: {
+                      topLeft: topLeft && isTop ? roundedCorner : 0,
+                      topRight: topRight && isTop ? roundedCorner : 0,
+                      bottomLeft: bottomLeft && isBottom ? roundedCorner : 0,
+                      bottomRight: bottomRight && isBottom ? roundedCorner : 0,
+                    },
+                  })}
+                />
+              );
+            }}
+          </CartesianChart>
+        </View>
 
         <ScrollView
           style={styles.optionsScrollView}
@@ -145,6 +209,40 @@ export default function NegativeBarChartsPage(props: { segment: string }) {
               title="Shuffle Data"
             />
           </View>
+          <Text style={{ fontWeight: "bold" }}>Rounded corners</Text>
+          <View
+            style={{
+              flexDirection: "row",
+              gap: 12,
+              marginTop: 10,
+              marginBottom: 16,
+            }}
+          >
+            <View style={{ gap: 10 }}>
+              <Checkbox
+                label={"Top left"}
+                checked={topLeft}
+                onChange={() => setTopLeft(!topLeft)}
+              />
+              <Checkbox
+                label={"Bottom left"}
+                checked={bottomLeft}
+                onChange={() => setBottomLeft(!bottomLeft)}
+              />
+            </View>
+            <View style={{ gap: 10 }}>
+              <Checkbox
+                label={"Top right"}
+                checked={topRight}
+                onChange={() => setTopRight(!topRight)}
+              />
+              <Checkbox
+                label={"Bottom right"}
+                checked={bottomRight}
+                onChange={() => setBottomRight(!bottomRight)}
+              />
+            </View>
+          </View>
         </ScrollView>
       </SafeAreaView>
     </>
@@ -160,7 +258,7 @@ const styles = StyleSheet.create({
     },
   },
   chart: {
-    flex: 1.5,
+    flex: 1,
   },
   optionsScrollView: {
     flex: 1,