+++ lib/src/cartesian/hooks/useStackedBarPaths.ts
@@ -46,6 +46,7 @@ type Props = {
     rowIndex: number;
   }) => CustomizablePathProps & { roundedCorners?: RoundedCorners };
 };
+
 export const useStackedBarPaths = ({
   points,
   chartBounds,
@@ -68,49 +69,44 @@ export const useStackedBarPaths = ({
   // so that we know where to start drawing the next bar for each x value
   const barYPositionOffsetTracker = points.reduce(
     (acc, points) => {
-      points.map((point) => {
-        const xValue = point.xValue;
-        if (acc[xValue]) {
-          acc[xValue] += 0 as number;
-        } else {
-          acc[xValue] = 0 as number;
-        }
-      });
-
+      points.map((point) => (acc[point.xValue] = [0, 0]));
       return acc;
     },
-    {} as Record<InputFieldType, number>,
+    {} as Record<InputFieldType, number[]>,
   );
 
   const paths = React.useMemo(() => {
     const bars: StackedBarPath[] = [];
+    const xToBottomTopIndexMap = getXToPointIndexMap(points);
 
     points.forEach((pointsArray, i) => {
-      const isTop = i === points.length - 1; // the top "row" of the stack of bars
-      const isBottom = i === 0; // the bottom "row" of the stack bars
-
       pointsArray.forEach((point, j) => {
+        const isBottom = xToBottomTopIndexMap.get(point.x)?.[0] === i;
+        const isTop = xToBottomTopIndexMap.get(point.x)?.[1] === i;
         const { yValue, x, y } = point;
         if (typeof y !== "number") return;
+        const isPositive = (yValue ?? 0) > 0;
 
         // call for any additional bar options per bar
         const options = barOptions({
           columnIndex: i,
           rowIndex: j,
-          isBottom,
-          isTop,
+          isBottom: isPositive ? isBottom : isTop,
+          isTop: isPositive ? isTop : isBottom,
         });
         const { roundedCorners, color, ...ops } = options;
 
         const path = Skia.Path.Make();
-
         const barHeight = yScale(0) - y;
-        const offset = barYPositionOffsetTracker?.[point.xValue!] ?? 0;
+
+        const offset = isPositive
+          ? barYPositionOffsetTracker?.[point.xValue!]?.[0]
+          : barYPositionOffsetTracker?.[point.xValue!]?.[1];
 
         if (roundedCorners) {
           const nonUniformRoundedRect = createRoundedRectPath(
             x,
-            y - offset,
+            y - (offset ?? 0),
             barWidth,
             barHeight,
             roundedCorners,
@@ -121,14 +117,20 @@ export const useStackedBarPaths = ({
           path.addRect(
             Skia.XYWHRect(
               point.x - barWidth / 2,
-              y - offset,
+              y - (offset ?? 0),
               barWidth,
               barHeight,
             ),
           );
         }
 
-        barYPositionOffsetTracker[point.xValue!] = barHeight + offset; // accumulate the heights as we loop
+        if (notNullAndUndefined(offset)) {
+          if (isPositive) {
+            barYPositionOffsetTracker[point.xValue!]![0] = barHeight + offset; // accumulate the positive heights as we loop
+          } else {
+            barYPositionOffsetTracker[point.xValue!]![1] = barHeight + offset; // accumulate the negative heights as we loop
+          }
+        }
 
         const bar = {
           path,
@@ -145,3 +147,33 @@ export const useStackedBarPaths = ({
 
   return paths;
 };
+
+/**
+ * Returns a map of x values to a two value array where the first number is the index of
+ * the bottom bar and the second is the index of the top bar.
+ */
+const getXToPointIndexMap = (
+  points: PointsArray[],
+): Map<number, [number, number]> => {
+  const xToIndexMap = new Map<number, [number, number]>();
+  points.forEach((pointsArray, i) => {
+    pointsArray.forEach(({ x, y, yValue }) => {
+      if (
+        notNullAndUndefined(y) &&
+        notNullAndUndefined(yValue) &&
+        yValue !== 0
+      ) {
+        const current = xToIndexMap.get(x);
+        if (!current) {
+          xToIndexMap.set(x, [i, i]);
+        } else {
+          yValue > 0 ? (current[1] = i) : (current[0] = i);
+        }
+      }
+    });
+  });
+  return xToIndexMap;
+};
+
+const notNullAndUndefined = <T>(value: T | null | undefined): value is T =>
+  value !== null && value !== undefined;