+++ lib/src/types.ts
@@ -64,7 +64,9 @@ export type CartesianChartRenderArg<
   YK extends keyof NumericalFields<RawData>,
 > = {
   xScale: Scale;
+  xTicks: number[];
   yScale: Scale;
+  yTicks: number[];
   chartBounds: ChartBounds;
   canvasSize: { width: number; height: number };
   points: {
@@ -239,7 +241,7 @@ export type YAxisProps<
 };
 
 export type FrameInputProps = {
-  lineWidth?: number;
+  lineWidth?: SidedNumber;
   lineColor?: Color;
   linePathEffect?: DashPathEffectComponent;
 };